% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_boxscores.R
\name{create_table_boxscores}
\alias{create_table_boxscores}
\title{Create DunkDB Table: boxscores fact}
\usage{
create_table_boxscores(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: boxscores fact
}
