% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_coaches.R
\name{create_table_coaches}
\alias{create_table_coaches}
\title{Create DunkDB Table: coaches dimension}
\usage{
create_table_coaches(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: coaches dimension
}
