% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_event_tags.R
\name{create_table_event_tags}
\alias{create_table_event_tags}
\title{Create DunkDB Table: event_tags dimension}
\usage{
create_table_event_tags(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: event_tags dimension
}
