% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_events_mapping.R
\name{create_table_events_mapping}
\alias{create_table_events_mapping}
\title{Create DunkDB Table: events_mapping dimension}
\usage{
create_table_events_mapping(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: events_mapping dimension
}
