% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_table_games}
\alias{create_table_games}
\title{Create DunkDB Table: games fact}
\usage{
create_table_games(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: games fact
}
