% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_pbp.R
\name{create_table_pbp}
\alias{create_table_pbp}
\title{Create DunkDB Table: play-by-play dimension}
\usage{
create_table_pbp(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: play-by-play dimension
}
