% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_schools.R
\name{create_table_schools}
\alias{create_table_schools}
\title{Create DunkDB Table: schools dimension}
\usage{
create_table_schools(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: schools dimension
}
