% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_seasons.R
\name{create_table_seasons}
\alias{create_table_seasons}
\title{Create DunkDB Table: seasons dimension}
\usage{
create_table_seasons(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Table: seasons dimension
}
