% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_tables}
\alias{create_tables}
\title{Create DunkDB Database Tables}
\usage{
create_tables(con)
}
\arguments{
\item{con}{A DBI database connection object to the dunkdb DuckDB database.}
}
\value{
N/A. Executes SQL to create database table.
}
\description{
Create DunkDB Database Tables
}
