% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_temp_to_perm}
\alias{db_temp_to_perm}
\title{Insert Data From Temp to Permanent Table}
\usage{
db_temp_to_perm(con, temp_name, perm_name, overwrite = FALSE)
}
\arguments{
\item{con}{dunkdb database connection}

\item{temp_name}{Name of temp table}

\item{perm_name}{Name of permanent table}

\item{overwrite}{Default false. True if existing data should be removed.}
}
\value{
N/A
}
\description{
Insert Data From Temp to Permanent Table
}
