% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_dunkdb.R
\name{init_dunkdb}
\alias{init_dunkdb}
\title{Initialize DunkDB}
\usage{
init_dunkdb(path = getwd(), db_name = "dunkdb.ddb")
}
\arguments{
\item{path}{File path for database location.}

\item{db_name}{Name of DuckDB database.}
}
\value{
A DuckDB database connection.
}
\description{
\emph{dunkdb}() creates a DuckDB database and the necessary database objects
to store NCAA men's basketball data in a normalized relational model.
}
