% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data.R
\name{insert_data}
\alias{insert_data}
\title{Insert Data Into DunkDB}
\usage{
insert_data(con)
}
\arguments{
\item{con}{A database connection to DunkDB database.}
}
\value{
N/A. Inserts data but doesn't return anything.
}
\description{
Gets data from stats.ncaa.org and inserts into DunkDB database.
}
