% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data.R
\name{populate_name_mapping}
\alias{populate_name_mapping}
\title{Populate name_mapping table}
\usage{
populate_name_mapping(con)
}
\arguments{
\item{con}{A database connection to DunkDB database.}
}
\value{
N/A
}
\description{
Inserts a name-to-nickname \href{https://github.com/carltonnorthern/nicknames/raw/refs/heads/master/names.csv}{table}
into database. This is for standardizing player names as play-by-play data
does not consistently use same first name.
}
