% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data.R
\name{populate_seasons}
\alias{populate_seasons}
\title{Insert Data for seasons Dimensional Table}
\usage{
populate_seasons(con)
}
\arguments{
\item{con}{Dunkdb database connection}
}
\value{
N/A
}
\description{
Insert Data for seasons Dimensional Table
}
